PointerQ's CommitRead/Write() take no arguments
Previously the caller needed to indicate how many pointers had been
committed. Now we store the size of the last buffer inside the queue
itself and use this when committing reads and writes. This reduces some
flexibility as it was previously possible for the client code to use
only some of a buffer and commit only the part they had used. It seems
unlikely this flexibility was very useful in practice, and made client
code significantly more complex.

During this change the min(...) and powerOfTwo(...) utility methods were
moved out of the spscq/ package and into the flib/fmath/ package.