Recording read/write failures
Every time a queue prevents a read or a write operation that failure is
recorded in a readFail or writeFail field. This allows us to measure
which end of the queue is outpacing the other. This feature is primarily
to give insight into the construction of benchmarks which should
exercise scenarios where the producer or the consumer is faster.

It's worth noting that keeping all of the queues aligned for common
features like this is very awkward. Some mitigation of this should be
investigated, code generation or struct delegation.