Deleted ByteQ and UnsafeByteQ
These were the queues which provided us with the ability to send
arbitrary numbers of bytes through the queue, in contrast with
ByteChunkQ which has fixed sized messages.

These two queues were deleted to reduce the maintenance burden. If we
find a real demand for a ByteQ in the future it will be very easy to
resurrect it.