Added an interface for each queue type
The interfaces are intended primarily as succinct documentation for each
queue. Because of the extensive delegation of methods down to commonQ it
was not obvious which methods were provided by each queue. Now this
information can be easily understood by reading the interface.

No performance tests use the interfaces, they use raw structs directly.