spscq.newCommonQ(...) errors when size > 2^42
newCommonQ(...) now has a size limit enforced. No queue may be created
with size greater than 2^42. If the underlying ring-buffer is storing
bytes a 4,096 gigabyte ring-buffer can be created, which is enough for
most applications.

We chose this limit because we found that including powers of two up to
2^63 broke our tests. The tests assume that they are able to increment
the write/read indices past the size value. When size is large the
write/read indices will overflow causing tests to fail.

The solution is to impose a reasonable limit on size in the tests, but
it would seem unwise to allow users to construct instances of commonQ we
were explicitly unwilling to test, so the limit is imposed in the
newCommonQ(int64) constructor function.

The, perhaps conservative, limit of 2^42 can be raised in the future
without breaking backwards compatibility.