Added fmath.Max(int64,int64) int64
Including tests which similar to the fmath.Min(...) function.

The limits to the size of the random numbers generated has been increased
to 2^42. This is the same as the current limit on the size of ring-buffers
used in the queue/spscq packages. These limits are not necessary and the
tests should be expanded to cover a wider range of integer values.