Replaced commonQ tests with invariant tests
Previously we had a number of tests ensuring that specific properties
are maintained for a particular field, such as write or read. These have
been replaced with a collection of tests which compare the state change
that occurred with a method call.

Currently we have state change tests for the
AcquireWrite/ReleaseWrite/Lazy and AcquireRead/ReleaseRead/Lazy methods.