Adding ftime package
This package contains three functions.

Counter() returns integer values which are guaranteed to increase
monotonically across calls on the same CPU core. Counter() is intended
to be used as a fine resolution performance measurement tool. The value
returned by Counter() need not correspond to a particular unit of time,
e.g. nanoseconds. It is more common for the value to reflect completed
CPU clock cycles.

If IsCounterSteady() returns true then we guarantee that values returned
by Counter() increase at a uniform 'steady' rate, independent of the
clockrate of the CPU it is running on.

If IsCounterSMPMonotonic() returns true we guarantee that for two calls
to Counter(), A and B, where A 'happens before' A, the value of B will
always be greater than A, even if those two calls are on separate cores
on the same socket. We make no guarantees across CPU sockets.