Adding ReadSingleBlocking/WriteSingleBlocking(...)
PointerQ now supports blocking calls to single reads and writes. While
using these queues in the matching engine it was found that having
blocking calls available made the queues significantly easier to work
with.

In the future we will review whether or not to keep these methods and/or
add similar blocking versions to each of the other queues. If blocking
methods, which look very desirable, are added other things will need to
be prunced to maintain a reasonable interface surface area. Likely
candidates for sacrifice include the *Lazy methods and perhaps the
ByteChunkQ struct.