Test for reads/writes size broken at end of buffer
A call to commonQ.acquireRead/acquireWrite(...) can return less than the
requested buffer size if a buffer of the requested size would run past
the end of the ring-buffer.

i.e. (from % size) + bufferSize > size

The tests failed to account for this condition would report that the
'Actual read/write size could have been biggger...'

This is now fixed.