Concurrent tests breaks single producer/consumer
The tests would run a large number of independent reading/writing
goroutines causing test failures in a scenario which the queues were
never intended to run.

The full set of reads and writes are completed by a pair of goroutines
and each test run is synchronized by a boolean channel signalling
completion.