ASM for LazyStore replaced with Go code
You can see the trick used to implement this in runtime/atomic_amd64x.go
(which should appear in Go 1.5). The use of a a nop() function prevents
the surrounding function from being inlined by the compiler. Because the
functions are not inlined they cannot be reorded. Just what the doctor
ordered for LazyStore :)