numeric comparison functions which return 0 or 1
This replicates the behaviour found in C where boolean values were just
integer values with a fancy type declaration. This means that integer
comparisons can be used as numerical values without, potentially
expensive, branches. An example usage is to select elements in a size 2
array without using if statements.

These functions are a wee bit niche.