Adding build constraints to fatomic/lazy.go
The fatomic.LazyStore(...) function relies on the x86/amd64 hardware
memory model for safety. Specifically LazyStore requires that no
previous stores be reordered after this store. This guarantee is given
by the x86/amd64 memory model and the (shaky) assurance that the
LazyStore function will never be inlined into another function.

Although it has been noted before it is worth noting again that the use
of the nop() function here ensures that the compiler won't inline the
LazyStore function. This heavily depends on specific compiler
optimisations and the go version compiler constraints reflect this.