spscq perf - invalid pointer values caught in 1.7
Fixes github issue #2

This panic is caused by the 'trick' of creating `unsafe.Pointer` values
without allocating.

    unsafe.Pointer(uintptr(uint(t)))

This converts a simple integer `t` into a pointer. For each test run `t`
starts at 0 and is incremented in the test loop to provide a predictable
`unsafe.Pointer` that can be written to the queue.

Golang's garbage collector requires a write barrier for every write of a
pointer value. In the problematic test we can see

    buffer[i] = unsafe.Pointer(uintptr(uint(t)))

which is writing a pointer into the buffer, invoking the write barrier.

As of 1.7 the write barrier checks that the pointer is within a valid
range of pointer values specifically it checks whether the value is too
small (`<minPhysPageSize`). Our low `t` values are caught by this check
causing the panic we are seeing.

The solution to this immediate problem is to add an offset to `t` to
push it up into a valid range. For this we have added a

    func getValidPointer() uintptr {
            intVal := 0
            return uintptr(unsafe.Pointer(&intVal))
    }

function which creates a real pointer. We keep the same `t` values but
offset them with the `uintptr` generated here.

This solves our mbarrier problem, but we still trip up on new safety
checks within the 1.7 garbage collector. This is solved in a separate
commit.