perf_spscq allocates for all unsafe.Pointers
Fixes #3

Previously we were creating unsafe.Pointers for pointerq tests. We were
not actually allocating memory for these pointers, we were simply
incrementing an integer and casting it to unsafe.Pointer.

As of Go 1.7 the garbage collector notices when it its sweep phase
detects more memory allocations found than it expected. Specifically we
see

runtime: nelems=73 nfree=0 nalloc=73 previous allocCount=1 nfreed=65464
    fatal error: sweep increased allocation count

    runtime stack:
    runtime.throw(0xd750a, 0x20)
            /Users/fmstephe/golang/go/src/runtime/panic.go:566 +0x95
    runtime.(*mspan).sweep(0x18cbe0, 0x1100000000, 0xc400000001)
    ...

The temporary fix was to disable the garbage collector. But it was clear
that continuing to write and read pointers to unallocated addresses
would leave us open to further problems.

We now allocate integer values using the

    getValidPointers(num int64) (ptrs []unsafe.Pointer, checksum int64)

function. This function ensures that every unsafe.Pointer points to
allocated memory and also precomputes the checksum (int64 sum of all
pointers) so we don't need to perform that task in the consumer anymore.

Garbage collection remains disabled for perf tests.