PointerQ now uses Acquire/Release api
Previously PointerQ batched reads/writes were performed by passing a
slice argument which was then copied from/to the internal ring buffer.
PointerQ now mimics the Acquire/Release structure of ByteChunkQ and
ByteMsgQ. This directly exposes parts of the queue's internal ring
buffer eliminating unnecessary copying, but introducing some correctness
hazards.

The new interface looks like:

	AcquireWrite(int64) []unsafe.Pointer
	ReleaseWrite()

To perform a write the caller calls 'q.AcquireWrite(size)' and a buffer
'len(buffer) == size' or 'nil' is returned. After the buffer contents
have been written the caller must call 'q.ReleaseWrite()'. This allows
the queue to reuse the space in buffer.

	AcquireRead(int64) []unsafe.Pointer
	ReleaseRead()

To perform a read the caller calls 'q.AcquireRead(size)' and a buffer
'len(buffer) <= size' or nil is returned. After the buffer contents have
been read the caller calls 'q.ReleaseRead()'.

In contrast to 'AcquireWrite(size)` the length of the buffer returned by
`AcquireRead(size) may be anything from '0' to 'size', this design was
chosen because if the number of readable entries is less than 'size'
then they would be unreadable unless we allow for smaller buffers to be
returned. This is particularly awkward when the writer has shut down and
the reader wishes to drain the queue.

It is important to note that `q.AcquireRead/Write(...)' calls return a
sub-slice of the queue's internal ring-buffer. This means that the buffer
must only be used temporarily for reading, or writing and must never be
stored past the lifetime of the acquire/release method calls.