PointerQ now nils pointers after they are read
Previously we were simply leaving them in the ringbuffer and allowing
them to be overwritten. This is a clear potential memory leak.

As an aside, it is interesting to note that the speed at which these
queues can read and write points is much faster than the rate at which
Go can allocate on the heap. This suggests that PointerQ itself may not
be ot a useful tool, since sending unrecycled pointers is inherently
slow, regardless of how the queue is implemented. Sending recycled
pointers, via another queue or some kind of pool is possible, but
doesn't feel easy.

This thought suggests a new kind of queue, where the elements are
pre-populated in the ring-buffer and read/writes simply cycle through
these elements. This requires a data-copying approach where writes copy
values into pre-populated elements and readers copy the data back out.
This looks very similar to the '[]byte' queues we already have.